+++ 2-platforms/craft/craft/config/general.php
@@ -19,8 +19,11 @@ return [
         'cacheDuration' => 'P2W', // Set default template caching time to 2 weeks
         'generateTransformsBeforePageLoad' => true, // Generate Image Transforms before page load. Else caching may be a problem
         'extraAllowedFileExtensions' => 'json', // Allow upload of Json files, also used for external Video Plugin
-        'sendPoweredByHeader' => false,
-        'defaultWeekStartDay' => 1,
+        'sendPoweredByHeader' => false, // Disable sending a "Powered by Craft" header
+        'defaultWeekStartDay' => 1, // Start Week on Monday
+        'allowAutoUpdates' => (getenv('APP_ENV') == 'local') ? 'minor-only' : false, // Disable auto-updates on servers
+        'timezone' => 'Europe/Zurich', // Set Timezone to Zurich
+
 
         'devMode' => (getenv('APP_DEBUG')=='true'),
         'enableTemplateCaching' => (getenv('TEMPLATE_CACHING')=='true'),